gpio/sodaville: Mark broken due to core irqdomain migration
The sodaville driver doesn't build anymore due to the transition to
common irq_domain in the core code.  It needs to be reworked, but
the rework isn't trivial.  Since this is a new driver anyway for
v3.4, mark it as broken now and a fixup patch can re-enable it when
the rework change has been tested.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>