UPSTREAM: ARM: 7709/1: mcpm: Add explicit AFLAGS to support v6/v7 multiplatform kernels
The full mcpm layer is not likely to be relevant to v6 based
platforms, so a multiplatform kernel won't use that code if booted
on v6 hardware.

This patch modifies the AFLAGS for affected mcpm .S files to
specify armv7-a explicitly for that code.

Signed-off-by: Dave Martin <dave.martin@linaro.org>
Acked-by: Arnd Bergmann <arnd@arndb.de>
Acked-by: Nicolas Pitre <nico@linaro.org>
Signed-off-by: Russell King <rmk+kernel@arm.linux.org.uk>
(cherry picked from linux-next commit
	70100a02268423b7bceafdf0ed6b0245f9baf7e8)

BUG=chrome-os-partner:19820
TEST=Booted peach

Change-Id: Icfa095ba1a80bb5e8b8bf1c4c7f096016eaa211e
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59029

Conflicts:
	arch/arm/common/Makefile