UPSTREAM: ARM: mcpm: use -st dsb option prior to sev instructions
In a similar manner to our spinlock implementation, mcpm uses sev to
wake up cores waiting on a lock when the lock is unlocked. In order to
ensure that the final write unlocking the lock is visible, a dsb
instruction is executed immediately prior to the sev.

This patch changes these dsbs to use the -st option, since we only
require that the store unlocking the lock is made visible.

Acked-by: Nicolas Pitre <nico@linaro.org>
Reviewed-by: Dave Martin <dave.martin@arm.com>
Reviewed-by: Catalin Marinas <catalin.marinas@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>

(cherry-picked from commit 40a5c0b415f080638a744177653aac4527002bbf)

BUG=chrome-os-partner:24078
TEST=Run WebGL and video aging test.
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>

Change-Id: I494004b89731b6c608b36400563f3c2f4e7ac984
Reviewed-on: https://chromium-review.googlesource.com/177012
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Arjun K V <arjun.kv@samsung.com>