arm: exynos: don't disable KFC ACE and ACP port before L2 powerdown
Set SKIP_DEACTIVATE_ACEACP_IN_PWDN for the KFC cluster so that
the ACE and ACP port is not turned off before the L2 cache is
powered down.

BUG=chrome-os-partner:19820
TEST=Once IKS support is added and eanbled, boot on peach and check
status of the cores using:
cat /dev/bL_status
        0 1 2 3 L2 CCI
[A15]   0 1 0 1  1  1
[A7]    1 0 1 0  1  1

Load/unload the CPUs using a CPU loading application and ensure
core switching is occuring.

Change-Id: Ie52abddb7fb7edf1d6e66ed65b2dc802327e99ba
Signed-off-by: Thomas Abraham <thomas.ab@samsung.com>
Signed-off-by: Inderpal Singh <inderpal.s@samsung.com>
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59060
Reviewed-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Vidir Alexander Jonsson <valex@valex.is>