arm: bL_switcher: don't enable bL switcher on systems without CCI
Do not enable the big.LITTLE switcher on systems that do not have an
ARM CCI (cache-coherent interconnect) present.  Since the CCI is used
to main cache coherency between multiple clusters and peripherals, it
is unlikely that a system without CCI would support big.LITTLE.

BUG=chrome-os-partner:19820
TEST=With CONFIG_BL_SWITCHER, observe that the bL switcher is initialized
on Pit but not on Snow.

Change-Id: Ie41d0396e15b93069bd3f18dc15261baa106f40c
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59850

Conflicts:
	arch/arm/common/bL_switcher.c