arm: mcpm: indicate if CPU was powered up due to switch
mcpm_cpu_powered_up() is called both during a CPU switch and in
secondary CPU booting.  Since, during a switch, mcpm_cpu_powered_up()
may be used to power down an outgoing CPU/cluster, indicate to
mcpm_cpu_powered_up() whether or not the incoming CPU was started
becuase of a switch or not.

BUG=chrome-os-partner:19820
TEST=Boot Pit.  Once Exynos MCPM and IKS support is added, observe
that all 8 CPUs are booted initially and, after the switcher is
initialized, clusters are powered off when possible.

Change-Id: I0dfac698d9a673964b632932c11549b4d1b61423
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61218
Reviewed-by: Simon Glass <sjg@chromium.org>